requirements.txt
astroid==2.1.0
autopep8==1.4.3
Django==2.1.7
djangorestframework==3.9.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nanoid==2.0.0
pycodestyle==2.5.0
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pytz==2018.9
six==1.12.0
wrapt==1.11.1
